Fixed few issuse with student id

+++ src/main/java/com/mpk/samples/SpringReactUI/controller/StudentController.java
@@ -27,7 +27,7 @@ public class StudentController {
 	}
 
 	@GetMapping("/students/{id}")
-	public Student findById(@PathVariable Long id) {
+	public Student findById(@PathVariable Integer id) {
 		return studentsClient.retrieveStudentById(id).getBody();
 	}
